moved test cases to new folder
