+++ test/smart-contract-tests/ind_bank.wasm
