+++ test/smart-contract-tests/initContract.test.ts
@@ -6,16 +6,15 @@ import {
     InitContractPayload,
     ParameterType,
     SMParameter,
-    SMStruct,
-} from '../src/types';
+} from '../../src/types';
 import * as ed from 'noble-ed25519';
-import { getAccountTransactionSignDigest } from '../src/serialization';
-import { getNodeClient } from './testHelpers';
-import { AccountAddress } from '../src/types/accountAddress';
-import { GtuAmount } from '../src/types/gtuAmount';
-import { TransactionExpiry } from '../src/types/transactionExpiry';
+import { getAccountTransactionSignDigest } from '../../src/serialization';
+import { getNodeClient } from '../testHelpers';
+import { AccountAddress } from '../../src/types/accountAddress';
+import { GtuAmount } from '../../src/types/gtuAmount';
+import { TransactionExpiry } from '../../src/types/transactionExpiry';
 import { Buffer } from 'buffer/';
-import { ModuleReference } from '../src/types/moduleReference';
+import { ModuleReference } from '../../src/types/moduleReference';
 const client = getNodeClient();
 const senderAccountAddress =
     '4ZJBYQbVp3zVZyjCXfZAAYBVkJMyVj8UKUNj9ox5YqTCBdBq2M';
@@ -35,36 +34,23 @@ test('init contract with the wrong private key', async () => {
         sender: new AccountAddress(senderAccountAddress),
     };
 
-    const contractName = 'INDBankStruct1';
+    const contractName = 'INDBank';
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const inputParams: SMParameter<SMStruct> = {
-        type: ParameterType.Array,
-        value: [
-            {
-                type: ParameterType.U8,
-                value: 26,
-            } as SMParameter<number>,
-            {
-                type: ParameterType.U8,
-                value: 27,
-            } as SMParameter<number>,
-            {
-                type: ParameterType.U8,
-                value: 51,
-            } as SMParameter<number>,
-        ],
+    const params: SMParameter<undefined> = {
+        type: ParameterType.NoParameters,
+        value: undefined,
     };
-    const baseEnergy = 300000n;
+    const maxContractExecutionEnergy = 300000n;
 
     const initModule: InitContractPayload = {
         amount: new GtuAmount(0n),
         moduleRef: new ModuleReference(
-            '684f75962a199809997581a7405c04f4603c51c20deb080bc0be795d09dcead4'
+            'e51d9f9329f103faa18b1c99335281204df9e3eec23d7138f69ddd17fd63e9d0'
         ),
         contractName: contractName,
-        parameter: inputParams,
-        maxContractExecutionEnergy: baseEnergy,
-    } as InitContractPayload;
+        parameter: params,
+        maxContractExecutionEnergy: maxContractExecutionEnergy,
+    };
 
     const initContractTransaction: AccountTransaction = {
         header: header,
@@ -84,7 +70,6 @@ test('init contract with the wrong private key', async () => {
 
     const result = await client.sendAccountTransaction(
         initContractTransaction,
-
         signatures
     );
