+++ test/smart-contract-tests/initContractWithParametersAddress.test.ts
@@ -6,20 +6,21 @@ import {
     InitContractPayload,
     ParameterType,
     SMParameter,
-} from '../src/types';
+    SMStruct,
+} from '../../src/types';
 import * as ed from 'noble-ed25519';
-import { getAccountTransactionSignDigest } from '../src/serialization';
-import { getNodeClient } from './testHelpers';
-import { AccountAddress } from '../src/types/accountAddress';
-import { GtuAmount } from '../src/types/gtuAmount';
-import { TransactionExpiry } from '../src/types/transactionExpiry';
+import { getAccountTransactionHash, getAccountTransactionSignDigest } from '../../src/serialization';
+import { getNodeClient } from '../testHelpers';
+import { AccountAddress } from '../../src/types/accountAddress';
+import { GtuAmount } from '../../src/types/gtuAmount';
+import { TransactionExpiry } from '../../src/types/transactionExpiry';
 import { Buffer } from 'buffer/';
-import { ModuleReference } from '../src/types/moduleReference';
+import { ModuleReference } from '../../src/types/moduleReference';
 const client = getNodeClient();
 const senderAccountAddress =
-    '4ZJBYQbVp3zVZyjCXfZAAYBVkJMyVj8UKUNj9ox5YqTCBdBq2M';
+    '3gLPtBSqSi7i7TEzDPpcpgD8zHiSbWEmn23QZH29A7hj4sMoL5';
 const wrongPrivateKey =
-    'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
+    '681de9a98d274b56eace2f86eb134bfc414f5c366022f281335be0b2d45a8988';
 // test case for init contract
 test('init contract with the wrong private key', async () => {
     const nextAccountNonce = await client.getNextAccountNonce(
@@ -34,23 +35,23 @@ test('init contract with the wrong private key', async () => {
         sender: new AccountAddress(senderAccountAddress),
     };
 
-    const contractName = 'INDBank';
+    const contractName = 'UserAddress';
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const params: SMParameter<undefined> = {
-        type: ParameterType.NoParameters,
-        value: undefined,
+    const inputParams: SMParameter<AccountAddress> = {
+        type: ParameterType.AccountAddress,
+        value: new AccountAddress(senderAccountAddress),
     };
-    const maxContractExecutionEnergy = 300000n;
+    const baseEnergy = 300000n;
 
     const initModule: InitContractPayload = {
         amount: new GtuAmount(0n),
         moduleRef: new ModuleReference(
-            'e51d9f9329f103faa18b1c99335281204df9e3eec23d7138f69ddd17fd63e9d0'
+            '3116f9792bf412e273d10a52bf42b78dddb9833d0c51c6cf79831f3b96866569'
         ),
         contractName: contractName,
-        parameter: params,
-        maxContractExecutionEnergy: maxContractExecutionEnergy,
-    };
+        parameter: inputParams,
+        maxContractExecutionEnergy: baseEnergy,
+    } as InitContractPayload;
 
     const initContractTransaction: AccountTransaction = {
         header: header,
@@ -70,8 +71,14 @@ test('init contract with the wrong private key', async () => {
 
     const result = await client.sendAccountTransaction(
         initContractTransaction,
+
         signatures
     );
 
+    const txHash = await getAccountTransactionHash(
+        initContractTransaction,
+        signatures
+    );
+    console.log(txHash);
     expect(result).toBeTruthy();
 }, 300000);
